+++ src/api/mod.rs
@@ -1,0 +1,11 @@
+pub use self::config::Config;
+pub use self::server::Server;
+
+pub mod config;
+#[allow(clippy::from_over_into)]
+mod http_error;
+#[allow(clippy::from_over_into)]
+pub mod response;
+mod routes;
+#[allow(clippy::module_inception)]
+pub mod server;
