+++ src/api/routes/mod.rs
@@ -1,0 +1,3 @@
+pub use check::check_ban;
+
+pub mod check;
