+++ src/config.rs
@@ -1,7 +1,7 @@
 use pepe_config::{ConfigError, FileFormat};
 use serde::{Deserialize, Serialize};
 
-use crate::server;
+use crate::api;
 use crate::telemetry;
 
 pub const DEFAULT_CONFIG: &str = include_str!("../config.yaml");
@@ -9,7 +9,7 @@ pub const DEFAULT_CONFIG: &str = include_str!("../config.yaml");
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Config {
     pub redis: pepe_config::redis::Config,
-    pub server: server::Config,
+    pub server: api::Config,
     pub telemetry: telemetry::Config,
     pub namespace: String,
     #[serde(default = "default_redis_query_timeout_secs")]
@@ -17,7 +17,7 @@ pub struct Config {
 }
 
 fn default_redis_query_timeout_secs() -> u64 {
-    return 5;
+    5
 }
 
 impl Config {
