+++ src/lib.rs
@@ -1,9 +1,7 @@
+pub mod api;
 pub mod ban_checker;
 pub mod config;
 pub mod errors;
-#[allow(clippy::from_over_into)]
-pub mod http_error;
 pub mod model;
 pub mod redis;
-pub mod server;
 pub mod telemetry;
