+++ src/main.rs
@@ -1,6 +1,6 @@
+use firewall_executor::api::Server;
 use firewall_executor::ban_checker::redis::RedisBanChecker;
 use firewall_executor::redis::get_pool;
-use firewall_executor::server::Server;
 use firewall_executor::{config, telemetry};
 
 #[tokio::main]
