--- src/server/mod.rs
@@ -1,8 +1,0 @@
-pub mod config;
-#[allow(clippy::from_over_into)]
-pub mod response;
-#[allow(clippy::module_inception)]
-pub mod server;
-
-pub use self::config::Config;
-pub use self::server::Server;
